+++ src/core/internal/iframe-form/index.ts
@@ -75,9 +75,9 @@ export class IFrameFormElement extends EventEmitter {
 
   private sensitive: boolean;
 
-  tableName: string;
+  tableName?: string;
 
-  fieldName: string;
+  fieldName?: string;
 
   iFrameName: string;
 
@@ -112,27 +112,27 @@ export class IFrameFormElement extends EventEmitter {
     const frameValues = name.split(':');
 
     const fieldType = frameValues[1];
-    const tempfield = atob(frameValues[2]);
+    // const tempfield = atob(frameValues[2]);
 
-    const removeAfter = tempfield.indexOf(':');
-    const field = removeAfter === -1 ? tempfield : tempfield.substring(0, removeAfter);
+    // const removeAfter = tempfield.indexOf(':');
+    // const field = removeAfter === -1 ? tempfield : tempfield.substring(0, removeAfter);
     // set frame name as frame type of the string besides : is number
-    const [tableName, fieldName] = [
-      field.substr(0, field.indexOf('.')),
-      field.substr(field.indexOf('.') + 1),
-    ];
+    // const [tableName, fieldName] = [
+    //   field.substr(0, field.indexOf('.')),
+    //   field.substr(field.indexOf('.') + 1),
+    // ];
     this.containerType = metaData.containerType;
     this.iFrameName = name;
     this.fieldType = fieldType;
 
-    this.tableName = tableName;
-    this.fieldName = fieldName;
+    // this.tableName = tableName;
+    // this.fieldName = fieldName;
     this.label = label;
     this.skyflowID = skyflowID;
 
     this.sensitive = ELEMENTS[this.fieldType].sensitive;
 
-    this.state.name = fieldName;
+    // this.state.name = fieldName;
 
     this.metaData = metaData;
     this.context = context;
@@ -282,7 +282,8 @@ export class IFrameFormElement extends EventEmitter {
     } else if (!this.getUnformattedValue() && !this.state.isEmpty) {
       this.state.isEmpty = true;
     }
-    if (valid && !this.doesClientHasError && this.validator(this.state.value)) {
+
+    if (this.validator(this.state.value) && !this.doesClientHasError && valid) {
       this.state.isValid = true;
       if (this.fieldType === ELEMENTS.EXPIRATION_MONTH.name) {
         //
@@ -400,6 +401,9 @@ export class IFrameFormElement extends EventEmitter {
     let resp = true;
     if (this.validations && this.validations.length) {
       for (let i = 0; i < this.validations.length; i += 1) {
+        if (this.fieldType === ElementType.CARD_NUMBER) {
+          value = value.replace(/[\s-]/g, '');
+        }
         switch (this.validations[i].type) {
           case ValidationRuleType.REGEX_MATCH_RULE:
             if (this.validations[i].params.regex) {
@@ -431,6 +435,9 @@ export class IFrameFormElement extends EventEmitter {
                   .document.getElementById(elementId) as HTMLInputElement;
                 if (collectInputElement) {
                   elementValue = collectInputElement.value;
+                  if (this.fieldType === ElementType.CARD_NUMBER) {
+                    elementValue = elementValue.replace(/[\s-]/g, '');
+                  }
                 }
               }
               if (elementValue !== value) {
@@ -480,6 +487,18 @@ export class IFrameFormElement extends EventEmitter {
           if (data.options.value !== undefined) {
             // for setting value
             this.setValue(<string | undefined>data.options.value);
+            if (
+              !(
+                this.fieldType === ELEMENTS.EXPIRATION_MONTH.name
+                || this.fieldType === ELEMENTS.EXPIRATION_DATE.name
+              )
+            ) {
+              bus.emit(ELEMENT_EVENTS_TO_IFRAME.INPUT_EVENT, {
+                name: this.iFrameName,
+                event: ELEMENT_EVENTS_TO_CLIENT.CHANGE,
+                value: this.getStatus(),
+              });
+            }
           } else if (
             data.options !== undefined
             && data.isSingleElementAPI === true
@@ -544,11 +563,24 @@ export class IFrameFormElement extends EventEmitter {
   };
 
   sendChangeStatus = (inputEvent: boolean = false) => {
-    bus.emit(ELEMENT_EVENTS_TO_IFRAME.INPUT_EVENT, {
-      name: this.iFrameName,
-      event: ELEMENT_EVENTS_TO_CLIENT.CHANGE,
-      value: this.getStatus(),
-    });
+    if (this.state.isFocused) {
+      bus.emit(ELEMENT_EVENTS_TO_IFRAME.INPUT_EVENT, {
+        name: this.iFrameName,
+        event: ELEMENT_EVENTS_TO_CLIENT.CHANGE,
+        value: this.getStatus(),
+      });
+    } else if (
+      this.state.value
+      && (this.fieldType === ELEMENTS.EXPIRATION_DATE.name
+        || this.fieldType === ELEMENTS.EXPIRATION_MONTH.name
+        || this.fieldType === ELEMENTS.FILE_INPUT.name)
+    ) {
+      bus.emit(ELEMENT_EVENTS_TO_IFRAME.INPUT_EVENT, {
+        name: this.iFrameName,
+        event: ELEMENT_EVENTS_TO_CLIENT.CHANGE,
+        value: this.getStatus(),
+      });
+    }
 
     this._emit(ELEMENT_EVENTS_TO_CLIENT.CHANGE, {
       ...this.getStatus(),
@@ -882,63 +914,65 @@ export class IFrameForm {
       const {
         state, tableName, validations, skyflowID,
       } = this.iFrameFormElements[formElements[i]];
-      if (
-        this.iFrameFormElements[formElements[i]].fieldType
-        !== ELEMENTS.FILE_INPUT.name
-      ) {
+      if (tableName) {
         if (
           this.iFrameFormElements[formElements[i]].fieldType
-          === ELEMENTS.checkbox.name
+        !== ELEMENTS.FILE_INPUT.name
         ) {
-          if (insertResponseObject[state.name]) {
-            insertResponseObject[state.name] = `${insertResponseObject[state.name]},${state.value
-            }`;
-          } else {
-            insertResponseObject[state.name] = state.value;
-          }
-        } else if (insertResponseObject[tableName] && (!skyflowID || skyflowID === undefined)) {
-          if (get(insertResponseObject[tableName], state.name)
+          if (
+            this.iFrameFormElements[formElements[i]].fieldType
+          === ELEMENTS.checkbox.name
+          ) {
+            if (insertResponseObject[state.name]) {
+              insertResponseObject[state.name] = `${insertResponseObject[state.name]},${state.value
+              }`;
+            } else {
+              insertResponseObject[state.name] = state.value;
+            }
+          } else if (insertResponseObject[tableName] && (!skyflowID || skyflowID === undefined)) {
+            if (get(insertResponseObject[tableName], state.name)
             && !(validations && checkForElementMatchRule(validations))) {
-            return Promise.reject(new SkyflowError(SKYFLOW_ERROR_CODE.DUPLICATE_ELEMENT,
-              [state.name, tableName], true));
-          }
-          set(
-            insertResponseObject[tableName],
-            state.name,
-            this.iFrameFormElements[formElements[i]].getUnformattedValue(),
-          );
-        } else if (skyflowID) {
-          if (skyflowID === null) {
-            return Promise.reject(new SkyflowError(
-              SKYFLOW_ERROR_CODE.EMPTY_SKYFLOW_ID_IN_ADDITIONAL_FIELDS,
-            ));
-          }
-          if (updateResponseObject[skyflowID]) {
+              return Promise.reject(new SkyflowError(SKYFLOW_ERROR_CODE.DUPLICATE_ELEMENT,
+                [state.name, tableName], true));
+            }
             set(
-              updateResponseObject[skyflowID],
+              insertResponseObject[tableName],
               state.name,
               this.iFrameFormElements[formElements[i]].getUnformattedValue(),
             );
+          } else if (skyflowID) {
+            if (skyflowID === null) {
+              return Promise.reject(new SkyflowError(
+                SKYFLOW_ERROR_CODE.EMPTY_SKYFLOW_ID_IN_ADDITIONAL_FIELDS,
+              ));
+            }
+            if (updateResponseObject[skyflowID]) {
+              set(
+                updateResponseObject[skyflowID],
+                state.name,
+                this.iFrameFormElements[formElements[i]].getUnformattedValue(),
+              );
+            } else {
+              updateResponseObject[skyflowID] = {};
+              set(
+                updateResponseObject[skyflowID],
+                state.name,
+                this.iFrameFormElements[formElements[i]].getUnformattedValue(),
+              );
+              set(
+                updateResponseObject[skyflowID],
+                'table',
+                tableName,
+              );
+            }
           } else {
-            updateResponseObject[skyflowID] = {};
+            insertResponseObject[tableName] = {};
             set(
-              updateResponseObject[skyflowID],
+              insertResponseObject[tableName],
               state.name,
               this.iFrameFormElements[formElements[i]].getUnformattedValue(),
             );
-            set(
-              updateResponseObject[skyflowID],
-              'table',
-              tableName,
-            );
           }
-        } else {
-          insertResponseObject[tableName] = {};
-          set(
-            insertResponseObject[tableName],
-            state.name,
-            this.iFrameFormElements[formElements[i]].getUnformattedValue(),
-          );
         }
       }
     }